Safari-611.1.10/LayoutTests/media/picture-in-picture/picture-in-picture-api-pip-window-expected.txt

This tests that a pip window is returned correctly when a video element enters the Picture-in-Picture mode.

RUN(internals.settings.setAllowsPictureInPictureMediaPlayback(true))
RUN(internals.setMockVideoPresentationModeEnabled(true))
RUN(video.src = findMediaFile("video", "../content/test"))
EVENT(canplaythrough)
EXPECTED (pipWindow.width > '0') OK
EXPECTED (pipWindow.height > '0') OK
END OF TEST
